更改秘法充能规则，屏幕适配

+++ BumperCars/Assets/Scripts/GameMain/Gameplay/UI/UltimateSkillBar.cs
@@ -9,7 +9,7 @@ namespace Lachesis.GamePlay
         public RectTransform barImageTrans;
         public int signelWidth = 110;
         
-        public void SetLevel(int level)
+        public void SetLevel(float level)
         {
             if(level>3) level = 3;
             Vector2 sizeDelta = barImageTrans.sizeDelta;
